% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sante.R
\name{anemie}
\alias{anemie}
\title{Calculer la prevalence de l'anemie}
\usage{
anemie(
  donnees,
  var_hemoglobine,
  var_groupe_cible = NULL,
  poids = NULL,
  sous_groupes = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees menages ou individus}

\item{var_hemoglobine}{character -- Taux d'hemoglobine en g/dL}

\item{var_groupe_cible}{character ou NULL -- Groupe cible :
"femmes_15_49", "enfants_6_59", "enceintes", "tous".
Defaut : NULL (tous)}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{sous_groupes}{character ou NULL -- Variables de desagregation.
Defaut : NULL}
}
\value{
Un tibble avec prevalence par niveau d'anemie
}
\description{
Calcule la prevalence de l'anemie selon les seuils OMS
(hemoglobine en g/dL). Distingue les niveaux legere, moderee, severe.
}
\examples{
set.seed(42)
n <- 400
femmes <- data.frame(
  hemoglobine = rnorm(n, 11.5, 2.1),
  milieu      = sample(c("urbain","rural"), n, TRUE),
  poids       = runif(n, 0.8, 1.3)
)
anemie(femmes, var_hemoglobine = "hemoglobine", poids = "poids")

}
