% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bien_etre.R
\name{bonheur_declare}
\alias{bonheur_declare}
\title{Calculer la part de population se declarant heureuse}
\usage{
bonheur_declare(
  donnees,
  var_bonheur,
  codes_heureux = c(1, 2),
  poids = NULL,
  sous_groupes = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees individuelles}

\item{var_bonheur}{character -- Variable de bonheur.
Valeurs attendues : 1="Tres heureux", 2="Heureux", 3="Pas tres heureux",
4="Pas du tout heureux" OU variable 0/1 deja binaire.}

\item{codes_heureux}{numeric ou character -- Codes consideres comme
"heureux" (les autres = "pas heureux"). Defaut : c(1, 2)}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{sous_groupes}{character ou NULL -- Variables de desagregation.
Defaut : NULL}
}
\value{
Un objet de classe \code{saf_subjectif}
}
\description{
Proportion d'individus se declarant tres heureux ou heureux.
Variable categorielle recodee en indicateur 0/1.
}
\examples{
set.seed(42)
n <- 500
individus <- data.frame(
  bonheur  = sample(1:4, n, TRUE, prob=c(0.20,0.45,0.25,0.10)),
  milieu   = sample(c("urbain","rural"), n, TRUE),
  poids    = runif(n, 0.8, 1.3)
)
bonheur_declare(individus, "bonheur", codes_heureux=c(1,2),
                 poids="poids", sous_groupes="milieu")

}
