% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inegalites.R
\name{calcul_gini}
\alias{calcul_gini}
\title{Calculer le coefficient de Gini}
\usage{
calcul_gini(
  donnees,
  var_revenu,
  poids = NULL,
  ic = FALSE,
  n_bootstrap = 200L,
  sous_groupes = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees menages}

\item{var_revenu}{character -- Variable de revenu ou de consommation
(valeurs strictement positives)}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{ic}{logical -- Calculer l'IC 95\% par bootstrap. Defaut : FALSE}

\item{n_bootstrap}{integer -- Replications bootstrap. Defaut : 200L}

\item{sous_groupes}{character ou NULL -- Variables de sous-groupe pour
decomposition. Defaut : NULL}
}
\value{
Un objet de classe \code{saf_gini}
}
\description{
Calcule le coefficient de Gini avec intervalle de confiance
par bootstrap. Le Gini mesure l'inegalite de distribution d'une variable
de revenu ou de consommation (0 = egalite parfaite, 1 = inegalite maximale).
}
\examples{
set.seed(42)
n <- 300
menages <- data.frame(
  depense = c(rexp(200, 1/150000), rexp(100, 1/500000)),
  milieu  = sample(c("urbain","rural"), n, TRUE),
  poids   = runif(n, 0.8, 1.3)
)
res <- calcul_gini(menages, "depense", poids = "poids")
print(res)

}
