% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographie.R
\name{carte_exporter}
\alias{carte_exporter}
\title{Exporter une carte en fichier image}
\usage{
carte_exporter(carte, chemin, largeur = 20, hauteur = 15, resolution = 300L)
}
\arguments{
\item{carte}{ggplot -- Objet ggplot2}

\item{chemin}{character -- Chemin de sortie (.png, .pdf, .svg)}

\item{largeur}{numeric -- Largeur en cm. Defaut : 20}

\item{hauteur}{numeric -- Hauteur en cm. Defaut : 15}

\item{resolution}{integer -- Resolution DPI (PNG). Defaut : 300L}
}
\value{
Chemin du fichier (invisible)
}
\description{
Exporte un objet ggplot2 en PNG, PDF ou SVG.
}
\examples{
\dontrun{
  g <- carte_choroplethe(sf_enrichi, "taux_pauvrete")
  carte_exporter(g, file.path(tempdir(), "carte.png"))
}

}
