% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{carte_ipm}
\alias{carte_ipm}
\title{Carte des privations IPM par zone}
\usage{
carte_ipm(
  decomp_ipm,
  sf_obj,
  cle_sf,
  cle_decomp,
  var = c("IPM", "H", "A"),
  titre = NULL,
  source = NULL
)
}
\arguments{
\item{decomp_ipm}{tibble -- Resultat de \code{decomposer_ipm()}}

\item{sf_obj}{sf -- Fond de carte (depuis \code{carte_zones()})}

\item{cle_sf}{character -- Variable cle dans \code{sf_obj}}

\item{cle_decomp}{character -- Variable cle dans \code{decomp_ipm}}

\item{var}{character -- Variable a cartographier :
\code{"IPM"}, \code{"H"} ou \code{"A"}. Defaut : "IPM"}

\item{titre}{character ou NULL -- Titre. Defaut : NULL}

\item{source}{character ou NULL -- Source. Defaut : NULL}
}
\value{
Un objet \code{ggplot2}
}
\description{
Cartographie l'IPM ou une composante (H, A, ou un
indicateur de privation) par zone geographique. Necessite un objet
sf et les resultats de decomposer_ipm().
}
\examples{
\dontrun{
  rca <- carte_zones("rca")
  res <- calcul_ipm(menages_rca, ...)
  decomp <- decomposer_ipm(res, menages_rca, "prefecture")
  carte_ipm(decomp, rca, cle_sf="prefecture",
            cle_decomp="prefecture")
}

}
