% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographie.R
\name{carte_pauvrete}
\alias{carte_pauvrete}
\title{Carte thematique de la pauvrete}
\usage{
carte_pauvrete(
  sf_obj,
  var_fgt0,
  seuil_alerte = 0.5,
  titre = NULL,
  source = NULL,
  afficher_valeurs = FALSE,
  var_label = NULL
)
}
\arguments{
\item{sf_obj}{sf -- Objet geographique enrichi avec les taux de pauvrete}

\item{var_fgt0}{character -- Variable du taux de pauvrete (FGT0, en
proportion ou pourcentage)}

\item{seuil_alerte}{numeric -- Seuil d'alerte (zones en rouge).
Defaut : 0.5 (50\%)}

\item{titre}{character ou NULL -- Titre. Defaut : titre automatique}

\item{source}{character ou NULL -- Note source. Defaut : NULL}

\item{afficher_valeurs}{logical -- Afficher les valeurs sur la carte.
Defaut : FALSE}

\item{var_label}{character ou NULL -- Variable a utiliser comme etiquette
(nom de la zone). Defaut : NULL}
}
\value{
Un objet \code{ggplot2}
}
\description{
Surcouche de \code{carte_choroplethe()} specialisee pour
la cartographie des indices de pauvrete (FGT0, FGT1, FGT2).
Utilise une symbologie standardisee AFRISTAT/Banque mondiale.
}
\examples{
\dontrun{
  carte_pauvrete(regions_enrichi,
                 var_fgt0 = "taux_pauvrete",
                 seuil_alerte = 0.5,
                 titre = "Incidence de la pauvrete par region 2026")
}

}
