% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sondage.R
\name{creer_design}
\alias{creer_design}
\title{Creer un objet design d'enquete complexe}
\usage{
creer_design(
  donnees,
  var_poids,
  var_strate = NULL,
  var_grappe = NULL,
  var_fpc = NULL,
  type = c("stratifie_grappes", "stratifie", "en_grappes", "simple")
)
}
\arguments{
\item{donnees}{data.frame -- Donnees de l'enquete}

\item{var_poids}{character -- Variable de poids de sondage}

\item{var_strate}{character ou NULL -- Variable de stratification.
Defaut : NULL}

\item{var_grappe}{character ou NULL -- Variable d'unites primaires
de sondage (grappes / UPS). Defaut : NULL}

\item{var_fpc}{character ou NULL -- Variable de correction de population
finie (taille de la strate dans la population). Defaut : NULL}

\item{type}{character -- Type de design : "stratifie", "en_grappes",
"stratifie_grappes", "simple". Defaut : "stratifie_grappes"}
}
\value{
Un objet de classe \code{saf_design} encapsulant un objet
\code{survey::svydesign}
}
\description{
Cree un objet de design d'enquete en precisant les strates,
grappes, poids de sondage et la correction de population finie (FPC).
Compatible avec les enquetes EHCVM, DHS, MICS, EFT, RGPH.
}
\examples{
set.seed(42)
n <- 300
donnees <- data.frame(
  poids_sond = runif(n, 1800, 3500),
  strate     = sample(c("Urbain","Rural","Semi-urbain"), n, TRUE),
  grappe     = sample(1:30, n, TRUE),
  revenu     = pmax(0, rnorm(n, 180000, 90000)),
  pauvre     = rbinom(n, 1, 0.45)
)
design <- creer_design(donnees, "poids_sond",
                        var_strate="strate", var_grappe="grappe")
print(design)

}
