% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sante.R
\name{emaciation}
\alias{emaciation}
\title{Calculer le taux d'emaciation (wasting)}
\usage{
emaciation(
  donnees,
  var_poids_taille_z,
  poids = NULL,
  sous_groupes = NULL,
  seuil = -2
)
}
\arguments{
\item{donnees}{data.frame -- Donnees enfants < 5 ans}

\item{var_poids_taille_z}{character -- Score Z poids-pour-taille (WHZ)}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{sous_groupes}{character ou NULL -- Variables de desagregation.
Defaut : NULL}

\item{seuil}{numeric -- Seuil score Z. Defaut : -2}
}
\value{
Un objet de classe \code{saf_anthropo}
}
\description{
Prevalence de l'emaciation : poids-pour-taille < -2
ecarts-types (standards OMS 2006). Indicateur de malnutrition aigue.
}
\examples{
set.seed(42)
n <- 400
enfants <- data.frame(
  whz  = rnorm(n, -0.8, 1.1),
  poids = runif(n, 0.8, 1.3)
)
emaciation(enfants, var_poids_taille_z = "whz", poids = "poids")

}
