% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pauvrete.R
\name{graphique_fgt}
\alias{graphique_fgt}
\title{Graphique des indices FGT}
\usage{
graphique_fgt(
  fgt_obj,
  type = c("barres", "indices"),
  variable = NULL,
  couleur = "#1B4965"
)
}
\arguments{
\item{fgt_obj}{objet \code{saf_fgt} -- Resultat de \code{calcul_fgt()}}

\item{type}{character -- \code{"barres"} ou \code{"indices"}.
Defaut : "barres"}

\item{variable}{character ou NULL -- Variable de sous-groupe a representer}

\item{couleur}{character -- Couleur principale. Defaut : \code{"#1B4965"}}
}
\value{
Objet \code{ggplot2}
}
\description{
Visualise les indices FGT nationaux et/ou par sous-groupe.
}
\examples{
\dontrun{
  set.seed(42)
  menages <- data.frame(
    depense_pc = c(rexp(70, 1/150000), rexp(30, 1/400000)),
    poids  = runif(100, 0.8, 1.2),
    milieu = sample(c("urbain", "rural"), 100, TRUE, c(0.4, 0.6))
  )
  fgt <- calcul_fgt(menages, "depense_pc", 220000,
                    poids = "poids", sous_groupes = "milieu")
  graphique_fgt(fgt, type = "barres", variable = "milieu")
}
}
