% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm.R
\name{graphique_ipm}
\alias{graphique_ipm}
\title{Graphique des contributions IPM}
\usage{
graphique_ipm(
  res_ipm,
  type = c("indicateurs", "dimensions"),
  titre = NULL,
  source = NULL
)
}
\arguments{
\item{res_ipm}{saf_ipm -- Resultat de \code{calcul_ipm()} ou
\code{calcul_ipm_national()}}

\item{type}{character -- \code{"indicateurs"} ou \code{"dimensions"}.
Defaut : "indicateurs"}

\item{titre}{character ou NULL -- Titre du graphique. Defaut : NULL}

\item{source}{character ou NULL -- Note de source. Defaut : NULL}
}
\value{
Un objet \code{ggplot2}
}
\description{
Produit un graphique en barres des contributions de chaque
indicateur et chaque dimension a l'IPM. Retourne un objet ggplot2.
}
\examples{
set.seed(42)
n <- 200
menages <- data.frame(
  nutrition   = rbinom(n,1,0.35), electricite = rbinom(n,1,0.60),
  eau         = rbinom(n,1,0.40), scolarisation = rbinom(n,1,0.30),
  poids       = runif(n,0.8,1.3)
)
res <- calcul_ipm(menages, var_nutrition="nutrition",
  var_electricite="electricite", var_eau="eau",
  var_scolarisation="scolarisation", poids="poids")
graphique_ipm(res)

}
