% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collecte.R
\name{import_sas}
\alias{import_sas}
\title{Importer un fichier SAS}
\usage{
import_sas(chemin, chemin_formats = NULL, encoding = "UTF-8", verbose = TRUE)
}
\arguments{
\item{chemin}{character — Chemin vers le fichier .sas7bdat}

\item{chemin_formats}{character ou NULL — Chemin vers le fichier de
formats .sas7bcat (optionnel). Défaut : NULL.}

\item{encoding}{character — Encodage. Défaut : "UTF-8".}

\item{verbose}{logical — Afficher les messages. Défaut : TRUE.}
}
\value{
Un tibble.
}
\description{
Importe un fichier SAS (.sas7bdat) ou un fichier de formats
SAS (.sas7bcat) avec préservation des labels.
}
\examples{
\donttest{
  donnees <- import_sas("data/enquete_emploi.sas7bdat")
}
}
