% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collecte.R
\name{import_spss}
\alias{import_spss}
\title{Importer un fichier SPSS}
\usage{
import_spss(
  chemin,
  garder_labels = TRUE,
  convertir_labels = FALSE,
  encoding = NULL,
  verbose = TRUE
)
}
\arguments{
\item{chemin}{character — Chemin vers le fichier .sav ou .zsav}

\item{garder_labels}{logical — Conserver les labels. Défaut : TRUE.}

\item{convertir_labels}{logical — Convertir en facteurs. Défaut : FALSE.}

\item{encoding}{character ou NULL — Encodage. Si NULL, détection auto.
Défaut : NULL.}

\item{verbose}{logical — Afficher les messages. Défaut : TRUE.}
}
\value{
Un tibble.
}
\description{
Importe un fichier SPSS (.sav ou .zsav) avec préservation
des labels de variables et de valeurs SPSS.
}
\examples{
\donttest{
  mics <- import_spss("data/mics6_enfants.sav")
}
}
