% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inegalites.R
\name{indice_atkinson}
\alias{indice_atkinson}
\alias{atkinson}
\title{Calculer l'indice d'Atkinson}
\usage{
indice_atkinson(donnees, var_revenu, poids = NULL, epsilon = 1)

atkinson(donnees, var_revenu, poids = NULL, epsilon = 1)
}
\arguments{
\item{donnees}{data.frame -- Donnees menages}

\item{var_revenu}{character -- Variable de revenu}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{epsilon}{numeric -- Parametre d'aversion a l'inegalite (>0).
Defaut : 1.0}
}
\value{
Un scalaire numerique (indice d'Atkinson)
}
\description{
L'indice d'Atkinson est une mesure d'inegalite parametrique
dont le parametre epsilon reflete l aversion a l'inegalite de la societe.
Epsilon = 0 : indifference ; Epsilon = 1 : forte aversion.
}
\examples{
set.seed(42)
menages <- data.frame(
  depense = c(rexp(200, 1/150000), rexp(100, 1/500000)),
  poids   = runif(300, 0.8, 1.3)
)
indice_atkinson(menages, "depense", poids = "poids", epsilon = 1)

}
