% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inegalites.R
\name{indice_palma}
\alias{indice_palma}
\alias{palma}
\title{Calculer l'indice de Palma}
\usage{
indice_palma(donnees, var_revenu, poids = NULL)

palma(donnees, var_revenu, poids = NULL)
}
\arguments{
\item{donnees}{data.frame -- Donnees menages}

\item{var_revenu}{character -- Variable de revenu}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}
}
\value{
Un scalaire numerique (indice de Palma)
}
\description{
L'indice de Palma est le rapport entre la part de revenu
des 10\% les plus riches et celle des 40\% les plus pauvres. Il est
plus sensible aux extremes que le coefficient de Gini.
}
\examples{
set.seed(42)
menages <- data.frame(
  depense = c(rexp(200, 1/150000), rexp(100, 1/500000)),
  poids   = runif(300, 0.8, 1.3)
)
indice_palma(menages, "depense", poids = "poids")

}
