% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sante.R
\name{insuffisance_ponderale}
\alias{insuffisance_ponderale}
\title{Calculer le taux d'insuffisance ponderale}
\usage{
insuffisance_ponderale(
  donnees,
  var_poids_age_z,
  poids = NULL,
  sous_groupes = NULL,
  seuil = -2
)
}
\arguments{
\item{donnees}{data.frame -- Donnees enfants < 5 ans}

\item{var_poids_age_z}{character -- Score Z poids-pour-age (WAZ)}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{sous_groupes}{character ou NULL -- Variables de desagregation.
Defaut : NULL}

\item{seuil}{numeric -- Seuil score Z. Defaut : -2}
}
\value{
Un objet de classe \code{saf_anthropo}
}
\description{
Prevalence de l'insuffisance ponderale : poids-pour-age
< -2 ecarts-types chez les enfants de moins de 5 ans.
}
\examples{
set.seed(42)
n <- 400
enfants <- data.frame(
  waz  = rnorm(n, -1.0, 1.2),
  poids = runif(n, 0.8, 1.3)
)
insuffisance_ponderale(enfants, var_poids_age_z = "waz", poids = "poids")

}
