% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapports.R
\name{lister_templates}
\alias{lister_templates}
\title{Lister les templates de rapport disponibles}
\usage{
lister_templates()
}
\value{
Un tibble avec : nom, description, parametres, format_sortie
}
\description{
Retourne la liste des templates R Markdown disponibles
dans statAfrikR avec leur description et parametres.
}
\examples{
lister_templates()

}
