% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre.R
\name{mariage_precoce}
\alias{mariage_precoce}
\title{Calculer le taux de mariage precoce}
\usage{
mariage_precoce(
  donnees,
  var_marie_avant_18,
  var_cohorte = NULL,
  poids = NULL,
  sous_groupes = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees femmes (15-49 ans)}

\item{var_marie_avant_18}{character -- Variable 0/1 : marie(e) avant 18 ans}

\item{var_cohorte}{character ou NULL -- Variable de cohorte ou groupe
d'age pour analyser les tendances. Defaut : NULL}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{sous_groupes}{character ou NULL -- Variables de desagregation.
Defaut : NULL}
}
\value{
Un objet de classe \code{saf_genre}
}
\description{
Proportion de femmes/hommes maries avant 18 ans (ODD 5.3.1)
avec desagregation et analyse des tendances par cohorte d'age.
}
\examples{
set.seed(42)
n <- 500
femmes <- data.frame(
  marie_18  = rbinom(n, 1, 0.42),
  cohorte   = sample(c("15-19","20-24","25-29","30-34","35-49"),
                      n, TRUE),
  milieu    = sample(c("urbain","rural"), n, TRUE),
  poids     = runif(n, 0.8, 1.3)
)
mariage_precoce(femmes, "marie_18",
                 var_cohorte = "cohorte",
                 poids = "poids", sous_groupes = "milieu")

}
