% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sante.R
\name{mortalite_5ans}
\alias{mortalite_5ans}
\title{Calculer le taux de mortalite des enfants de moins de 5 ans}
\usage{
mortalite_5ans(
  donnees,
  var_naissances_vivantes,
  var_deces_enfants,
  poids = NULL,
  sous_groupes = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees femmes en age de procrer (15-49 ans)
ou historique des naissances}

\item{var_naissances_vivantes}{character -- Nombre total de naissances
vivantes par femme}

\item{var_deces_enfants}{character -- Nombre de deces d'enfants < 5 ans
par femme}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{sous_groupes}{character ou NULL -- Variables de desagregation.
Defaut : NULL}
}
\value{
Un objet de classe \code{saf_mortalite}
}
\description{
Calcule le taux de mortalite infanto-juvenile (U5MR)
en pour 1 000 naissances vivantes. ODD 3.2.1.
}
\examples{
set.seed(42)
n <- 600
femmes <- data.frame(
  naissances_vivantes = rpois(n, 3.5),
  deces_enfants       = rbinom(n, 4, 0.08),
  milieu              = sample(c("urbain","rural"), n, TRUE),
  poids               = runif(n, 0.8, 1.3)
)
mortalite_5ans(femmes,
  var_naissances_vivantes = "naissances_vivantes",
  var_deces_enfants       = "deces_enfants",
  poids                   = "poids")

}
