% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre.R
\name{parite_education}
\alias{parite_education}
\title{Calculer l'indice de parite filles/garcons en education}
\usage{
parite_education(
  donnees,
  var_scolarise,
  var_sexe,
  poids = NULL,
  var_niveau = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees menages ou enfants}

\item{var_scolarise}{character -- Variable 0/1 : enfant scolarise}

\item{var_sexe}{character -- Variable sexe (H/F, M/F, 1/2, homme/femme)}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{var_niveau}{character ou NULL -- Niveau scolaire pour ISP par
niveau. Defaut : NULL}
}
\value{
Un objet de classe \code{saf_genre}
}
\description{
Calcule l'Indice de Parite entre les Sexes (ISP) pour les
taux de scolarisation bruts et nets. ODD 4.5.1.
ISP = valeur filles / valeur garcons. ISP = 1 = parite parfaite.
}
\examples{
set.seed(42)
n <- 600
enfants <- data.frame(
  scolarise = rbinom(n, 1, 0.72),
  sexe      = sample(c("H","F"), n, TRUE),
  niveau    = sample(c("primaire","secondaire"), n, TRUE),
  poids     = runif(n, 0.8, 1.3)
)
parite_education(enfants, "scolarise", "sexe",
                  poids="poids", var_niveau="niveau")

}
