% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bien_etre.R
\name{privations_ressenties}
\alias{privations_ressenties}
\title{Calculer le score de privations ressenties}
\usage{
privations_ressenties(
  donnees,
  vars_privations,
  poids = NULL,
  sous_groupes = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees menages}

\item{vars_privations}{named character -- Vecteur nomme : type de
privation -> variable 0/1. Ex : c(Nourriture="manque_nour",
Eau="manque_eau", Medicaments="manque_med")}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{sous_groupes}{character ou NULL -- Variables de desagregation.
Defaut : NULL}
}
\value{
Un tibble avec prevalence par type de privation
}
\description{
Calcule la proportion de menages ayant declare souffrir
de privations subjectives (manque de nourriture, eau, medicaments,
argent) au cours des 12 derniers mois. Complement a l'IPM.
}
\examples{
set.seed(42)
n <- 400
menages <- data.frame(
  manque_nour = rbinom(n, 1, 0.48),
  manque_eau  = rbinom(n, 1, 0.35),
  manque_med  = rbinom(n, 1, 0.58),
  manque_arg  = rbinom(n, 1, 0.62),
  milieu      = sample(c("urbain","rural"), n, TRUE),
  poids       = runif(n, 0.8, 1.3)
)
privations_ressenties(menages,
  vars_privations = c(Nourriture="manque_nour",
                      Eau="manque_eau",
                      Medicaments="manque_med",
                      Argent="manque_arg"),
  poids="poids", sous_groupes="milieu")

}
