% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation_geo.R
\name{rapport_harmonisation}
\alias{rapport_harmonisation}
\title{Produire un rapport d'harmonisation geographique}
\usage{
rapport_harmonisation(
  donnees,
  var_zone,
  reference,
  var_ref = "nom_local",
  pays = "Pays",
  annee = as.integer(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{donnees}{data.frame -- Donnees source}

\item{var_zone}{character -- Variable zone}

\item{reference}{data.frame ou saf_concordance -- Reference geographique}

\item{var_ref}{character -- Variable zone de reference.
Defaut : "nom_local"}

\item{pays}{character -- Nom du pays. Defaut : "Pays"}

\item{annee}{integer -- Annee. Defaut : annee courante}
}
\value{
Un tibble de rapport d'harmonisation
}
\description{
Genere un rapport complet sur l'harmonisation geographique :
zones appariees, non appariees, ambigues. Format institutionnel
conforme aux standards IHSN/PARIS21.
}
\examples{
donnees <- data.frame(
  region = c("Bangui","Ombella-Mpoko","Kemo","Zone inconnue"),
  stringsAsFactors = FALSE
)
ref <- data.frame(
  nom_local = c("Bangui","Ombella-M'Poko","Kemo","Bamingui-Bangoran"),
  stringsAsFactors = FALSE
)
rapport_harmonisation(donnees, "region", ref, pays="RCA", annee=2026L)

}
