% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saf_cameroun_departements}
\alias{saf_cameroun_departements}
\title{Departements du Cameroun}
\format{
Objet sf avec colonnes : departement, code, geometry.
}
\source{
Natural Earth -- statAfrikR Foundation
}
\usage{
saf_cameroun_departements
}
\description{
10 regions du Cameroun.
}
\examples{
cam <- carte_zones('subdivisions', pays = 'CMR')
}
\keyword{datasets}
