% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saf_cedeao}
\alias{saf_cedeao}
\title{Zone CEDEAO -- 15 pays membres}
\format{
Objet sf (sous-ensemble de saf_pays_afrique)
}
\source{
Natural Earth -- statAfrikR Foundation
}
\usage{
saf_cedeao
}
\description{
Communaute Economique des Etats de l'Afrique de l'Ouest.
}
\examples{
cedeao <- carte_zones('cedeao')
}
\keyword{datasets}
