% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saf_cemac}
\alias{saf_cemac}
\title{Zone CEMAC -- 6 pays membres}
\format{
Objet sf (sous-ensemble de saf_pays_afrique)
}
\source{
Natural Earth -- statAfrikR Foundation
}
\usage{
saf_cemac
}
\description{
Cameroun, Centrafrique, Congo, Gabon,
Guinee Equatoriale, Tchad.
}
\examples{
cemac <- carte_zones('cemac')
}
\keyword{datasets}
