% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saf_eau}
\alias{saf_eau}
\title{Zone EAC -- Afrique de l'Est}
\format{
Objet sf (sous-ensemble de saf_pays_afrique)
}
\source{
Natural Earth -- statAfrikR Foundation
}
\usage{
saf_eau
}
\description{
Communaute d'Afrique de l'Est.
}
\examples{
eau <- carte_zones('eau')
}
\keyword{datasets}
