% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saf_pays_afrique}
\alias{saf_pays_afrique}
\title{Pays africains -- fond de carte integre}
\format{
Objet sf avec colonnes : pays, pays_fr, iso2, iso3,
sous_region, geometry.
}
\source{
Natural Earth via rnaturalearth -- statAfrikR Foundation
}
\usage{
saf_pays_afrique
}
\description{
Fond de carte des 54 pays africains.
}
\examples{
afrique <- carte_zones('afrique')
}
\keyword{datasets}
