% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saf_rca_prefectures}
\alias{saf_rca_prefectures}
\title{Prefectures de la Republique Centrafricaine}
\format{
Objet sf avec colonnes : pays, iso3, prefecture, code, geometry.
}
\source{
Natural Earth -- statAfrikR Foundation
}
\usage{
saf_rca_prefectures
}
\description{
17 prefectures de la RCA avec codes administratifs.
}
\examples{
rca <- carte_zones('rca')
}
\keyword{datasets}
