% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saf_sadc}
\alias{saf_sadc}
\title{Zone SADC -- Afrique Australe}
\format{
Objet sf (sous-ensemble de saf_pays_afrique)
}
\source{
Natural Earth -- statAfrikR Foundation
}
\usage{
saf_sadc
}
\description{
Communaute de Developpement de l'Afrique Australe.
}
\examples{
sadc <- carte_zones('sadc')
}
\keyword{datasets}
