% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saf_senegal_regions}
\alias{saf_senegal_regions}
\title{Regions du Senegal}
\format{
Objet sf avec colonnes : region, code, geometry.
}
\source{
Natural Earth -- statAfrikR Foundation
}
\usage{
saf_senegal_regions
}
\description{
14 regions du Senegal.
}
\examples{
sen <- carte_zones('subdivisions', pays = 'SEN')
}
\keyword{datasets}
