% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saf_subdivisions_afrique}
\alias{saf_subdivisions_afrique}
\title{Subdivisions de niveau 1 -- tous les pays africains}
\format{
Objet sf avec colonnes : pays, iso3, subdivision,
code_admin, geometry.
}
\source{
Natural Earth -- statAfrikR Foundation
}
\usage{
saf_subdivisions_afrique
}
\description{
Prefectures, regions et provinces de 43 pays africains
(744 subdivisions + 17 prefectures RCA = 761 total).
}
\examples{
cam <- carte_zones('subdivisions', pays = 'CMR')
}
\keyword{datasets}
