% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonisation_geo.R
\name{table_concordance}
\alias{table_concordance}
\title{Creer une table de concordance entre nomenclatures geographiques}
\usage{
table_concordance(
  donnees,
  var_nom_local = "nom_zone",
  var_code_ins = NULL,
  var_code_iso = NULL,
  var_code_gadm = NULL,
  var_niveau = NULL,
  var_pays = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees de reference avec les noms/codes
des zones administratives}

\item{var_nom_local}{character -- Variable nom local (utilise par l'INS).
Defaut : "nom_zone"}

\item{var_code_ins}{character ou NULL -- Variable code INS.
Defaut : NULL}

\item{var_code_iso}{character ou NULL -- Variable code ISO 3166-2.
Defaut : NULL}

\item{var_code_gadm}{character ou NULL -- Variable code GADM.
Defaut : NULL}

\item{var_niveau}{character ou NULL -- Variable niveau administratif
(1 = region, 2 = departement, etc.). Defaut : NULL}

\item{var_pays}{character ou NULL -- Variable pays (code ISO 3).
Defaut : NULL}
}
\value{
Un objet de classe \code{saf_concordance}
}
\description{
Cree une table de correspondance entre les differentes
nomenclatures geographiques utilisees en Afrique : nomenclature locale
des INS, codes ISO 3166-2, codes GADM, codes EHCVM, et noms officiels.
Essentielle pour harmoniser les donnees entre plusieurs sources.
}
\examples{
zones_rca <- data.frame(
  nom_zone   = c("Bangui","Ombella-M'Poko","Bamingui-Bangoran",
                  "Mbomou","Haute-Kotto"),
  code_ins   = c("01","02","03","04","05"),
  code_iso   = c("CF-BGF","CF-MP","CF-BB","CF-MB","CF-HK"),
  code_gadm  = c("GADMa001","GADMa002","GADMa003","GADMa004","GADMa005"),
  niveau     = rep(1L, 5),
  pays       = rep("CAF", 5),
  stringsAsFactors = FALSE
)
table_concordance(zones_rca, var_nom_local="nom_zone",
                   var_code_ins="code_ins", var_code_iso="code_iso",
                   var_code_gadm="code_gadm", var_niveau="niveau",
                   var_pays="pays")

}
