% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emploi.R
\name{taux_activite}
\alias{taux_activite}
\title{Calculer le taux d'activite}
\usage{
taux_activite(
  donnees,
  var_actif,
  var_age = NULL,
  age_min = 15L,
  age_max = 64L,
  poids = NULL,
  sous_groupes = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees individuelles ou menages}

\item{var_actif}{character -- Variable 0/1 : actif (employe ou chomeur)}

\item{var_age}{character ou NULL -- Variable age pour filtrer la
population en age de travailler. Defaut : NULL}

\item{age_min}{integer -- Age minimum. Defaut : 15L}

\item{age_max}{integer -- Age maximum. Defaut : 64L}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{sous_groupes}{character ou NULL -- Variables de desagregation
(sexe, milieu, region). Defaut : NULL}
}
\value{
Un objet de classe \code{saf_emploi}
}
\description{
Calcule le taux d'activite selon la definition du Bureau
International du Travail (BIT) : proportion de la population en age
de travailler qui est active (employes + chomeurs). ODD 8.
}
\examples{
set.seed(42)
n <- 500
individus <- data.frame(
  actif  = rbinom(n, 1, 0.65),
  age    = sample(15:64, n, TRUE),
  sexe   = sample(c("H","F"), n, TRUE),
  milieu = sample(c("urbain","rural"), n, TRUE),
  poids  = runif(n, 0.8, 1.3)
)
taux_activite(individus, "actif", var_age="age", poids="poids",
              sous_groupes=c("sexe","milieu"))

}
