% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emploi.R
\name{taux_emploi}
\alias{taux_emploi}
\title{Calculer le taux d'emploi}
\usage{
taux_emploi(
  donnees,
  var_employe,
  var_age = NULL,
  age_min = 15L,
  age_max = 64L,
  poids = NULL,
  sous_groupes = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees individuelles}

\item{var_employe}{character -- Variable 0/1 : a un emploi}

\item{var_age}{character ou NULL -- Variable age. Defaut : NULL}

\item{age_min}{integer -- Age minimum. Defaut : 15L}

\item{age_max}{integer -- Age maximum. Defaut : 64L}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{sous_groupes}{character ou NULL -- Variables de desagregation.
Defaut : NULL}
}
\value{
Un objet de classe \code{saf_emploi}
}
\description{
Proportion de la population en age de travailler ayant
un emploi (salarie ou independant). Indicateur cle ODD 8.5.
}
\examples{
set.seed(42)
individus <- data.frame(
  employe = rbinom(500, 1, 0.55),
  age     = sample(15:64, 500, TRUE),
  sexe    = sample(c("H","F"), 500, TRUE),
  poids   = runif(500, 0.8, 1.3)
)
taux_emploi(individus, "employe", var_age="age",
            poids="poids", sous_groupes="sexe")

}
