% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emploi.R
\name{travail_enfants}
\alias{travail_enfants}
\title{Calculer la prevalence du travail des enfants}
\usage{
travail_enfants(
  donnees,
  var_travail_enfant,
  var_age = NULL,
  poids = NULL,
  sous_groupes = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees enfants 5-17 ans}

\item{var_travail_enfant}{character -- Variable 0/1 : enfant au travail}

\item{var_age}{character ou NULL -- Variable age pour filtrer 5-17 ans.
Defaut : NULL}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{sous_groupes}{character ou NULL -- Variables de desagregation.
Defaut : NULL}
}
\value{
Un objet de classe \code{saf_emploi}
}
\description{
Prevalence du travail des enfants (5-17 ans) selon la
definition OIT/MICS. ODD 8.7.1.
}
\examples{
set.seed(42)
n <- 600
enfants <- data.frame(
  travail = rbinom(n, 1, 0.22),
  age     = sample(5:17, n, TRUE),
  sexe    = sample(c("H","F"), n, TRUE),
  milieu  = sample(c("urbain","rural"), n, TRUE),
  poids   = runif(n, 0.8, 1.3)
)
travail_enfants(enfants, "travail", var_age="age",
                 poids="poids", sous_groupes=c("sexe","milieu"))

}
