% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sante.R
\name{vaccination}
\alias{vaccination}
\title{Calculer la couverture vaccinale}
\usage{
vaccination(
  donnees,
  vars_vaccins,
  poids = NULL,
  var_carte_sante = NULL,
  sous_groupes = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees enfants}

\item{vars_vaccins}{named character -- Vecteur nomme : nom du vaccin ->
nom de la variable (0/1). Ex: c(DTC3="dtc3", Rougeole="rougeole")}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{var_carte_sante}{character ou NULL -- Variable indiquant si
l'enfant a une carte de sante (pour ajustement). Defaut : NULL}

\item{sous_groupes}{character ou NULL -- Variables de desagregation.
Defaut : NULL}
}
\value{
Un tibble avec couverture par antigene
}
\description{
Calcule la couverture vaccinale par antigene selon la
methodologie DHS/MICS. Indicateurs ODD 3.b.1.
}
\examples{
set.seed(42)
n <- 500
enfants <- data.frame(
  dtc3      = rbinom(n, 1, 0.72),
  rougeole  = rbinom(n, 1, 0.68),
  polio3    = rbinom(n, 1, 0.75),
  bcg       = rbinom(n, 1, 0.90),
  poids     = runif(n, 0.8, 1.3),
  milieu    = sample(c("urbain","rural"), n, TRUE)
)
vaccination(enfants,
  vars_vaccins = c(DTC3="dtc3", Rougeole="rougeole",
                   Polio3="polio3", BCG="bcg"),
  poids = "poids")

}
