% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre.R
\name{violence_basee_genre}
\alias{violence_basee_genre}
\title{Calculer la prevalence des violences basees sur le genre}
\usage{
violence_basee_genre(
  donnees,
  var_violence_physique = NULL,
  var_violence_sexuelle = NULL,
  var_violence_psycho = NULL,
  poids = NULL,
  sous_groupes = NULL
)
}
\arguments{
\item{donnees}{data.frame -- Donnees femmes (15-49 ans)}

\item{var_violence_physique}{character ou NULL -- VBG physique (0/1)}

\item{var_violence_sexuelle}{character ou NULL -- VBG sexuelle (0/1)}

\item{var_violence_psycho}{character ou NULL -- VBG psychologique (0/1)}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}

\item{sous_groupes}{character ou NULL -- Variables de desagregation.
Defaut : NULL}
}
\value{
Un tibble avec prevalence par type de violence
}
\description{
Prevalence des violences physiques, sexuelles et/ou
psychologiques exercees par un partenaire intime ou un non-partenaire.
Methodologie DHS/OMS. ODD 5.2.1.
}
\examples{
set.seed(42)
n <- 500
femmes <- data.frame(
  vbg_phys   = rbinom(n, 1, 0.28),
  vbg_sex    = rbinom(n, 1, 0.18),
  vbg_psycho = rbinom(n, 1, 0.35),
  milieu     = sample(c("urbain","rural"), n, TRUE),
  poids      = runif(n, 0.8, 1.3)
)
violence_basee_genre(femmes,
  var_violence_physique  = "vbg_phys",
  var_violence_sexuelle  = "vbg_sex",
  var_violence_psycho    = "vbg_psycho",
  poids = "poids", sous_groupes = "milieu")

}
