% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographie.R
\name{whipple}
\alias{whipple}
\title{Calculer l'indice de Whipple}
\usage{
whipple(donnees, var_age, age_min = 23L, age_max = 62L, poids = NULL)
}
\arguments{
\item{donnees}{data.frame -- Donnees individuelles}

\item{var_age}{character -- Variable age en annees revolues}

\item{age_min}{integer -- Age minimum pour le calcul. Defaut : 23L}

\item{age_max}{integer -- Age maximum pour le calcul. Defaut : 62L}

\item{poids}{character ou NULL -- Variable de ponderation. Defaut : NULL}
}
\value{
Un objet de classe \code{saf_qualite_demo}
}
\description{
Mesure l'attraction des ages se terminant par 0 et 5 dans
les declarations d'age. Un indice de Whipple < 105 indique des donnees
de tres bonne qualite ; > 175 indique une qualite tres mauvaise.
Methodologie Nations Unies / IUSSP.
}
\examples{
set.seed(42)
n <- 1000
# Simulation d'attraction sur les multiples de 5
ages_base <- sample(23:62, n, TRUE)
attire    <- ages_base - (ages_base \%\% 5) # attire vers 0 et 5
ages      <- ifelse(runif(n) < 0.25, attire, ages_base)
individus <- data.frame(age = ages, poids = runif(n, 0.8, 1.3))
whipple(individus, "age")

}
