% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographie.R
\name{carte_import}
\alias{carte_import}
\title{Importer un fichier geographique}
\usage{
carte_import(chemin, crs = 4326L, couche = NULL, simplifier = FALSE)
}
\arguments{
\item{chemin}{character -- Chemin vers le fichier (.shp, .geojson, .gpkg)}

\item{crs}{integer -- Code EPSG cible. Defaut : 4326 (WGS 84)}

\item{couche}{character ou NULL -- Couche GeoPackage. Defaut : NULL}

\item{simplifier}{logical -- Simplifier la geometrie. Defaut : FALSE}
}
\value{
Un objet \code{sf}
}
\description{
Importe un fichier geographique externe (shapefile, GeoJSON,
GeoPackage) fourni par l'utilisateur ou l'INS.
}
\examples{
\dontrun{
  regions <- carte_import("data/regions_enquete.shp")
}

}
