% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{palette_ins}
\alias{palette_ins}
\title{Palette de couleurs INS}
\usage{
palette_ins(n = 6, type = c("categoriel", "sequentiel", "divergent"))
}
\arguments{
\item{n}{integer — Nombre de couleurs souhaité. Défaut : 6.}

\item{type}{character — \code{"categoriel"}, \code{"sequentiel"},
\code{"divergent"}. Défaut : "categoriel".}
}
\value{
Vecteur de codes couleurs hexadécimaux.
}
\description{
Retourne une palette de couleurs officielle pour les
graphiques INS, compatible daltonisme.
}
\examples{
\dontrun{
  couleurs <- palette_ins(4)
}
}
