% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statAfrikR-package.R
\docType{package}
\name{statAfrikR-package}
\alias{statAfrikR}
\alias{statAfrikR-package}
\title{statAfrikR — Outils statistiques pour les INS africains}
\description{
Boite a outils statistique complete pour les Instituts
Nationaux de Statistique (INS) d'Afrique.
}
\seealso{
Site web : \url{https://statafrikr.org/}
GitHub : \url{https://github.com/damoko2004/statAfrikR}
Discord : \url{https://discord.com/invite/kcfA27Yz}
}
\author{
\strong{Maintainer}: Dikers Amoko \email{diamoko@gmail.com} [contributor]

Other contributors:
\itemize{
  \item Josue Honore Dasse \email{josue.h.dasse@gmail.com} [contributor]
}

}
\keyword{internal}
