% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statAfrikR-package.R
\docType{package}
\name{statAfrikR-package}
\alias{statAfrikR}
\alias{statAfrikR-package}
\title{statAfrikR: Outils Statistiques pour les Instituts Nationaux de Statistique Africains}
\description{
Boîte à outils statistiques complète pour les Instituts Nationaux de Statistique (INS) africains. Fournit des fonctions pour l'import de données d'enquêtes (KoboToolbox, ODK, CSPro, Excel, Stata, SPSS), le traitement et la validation, l'analyse statistique pondérée (statistiques descriptives, tableaux croisés, régression, IDH, IPM, inégalités), la visualisation (pyramides des âges, cartes thématiques, graphiques officiels) et la diffusion (rapports Word/PDF, anonymisation, export SDMX, métadonnées DDI). Conçu pour fonctionner en environnement à ressources limitées, hors ligne et en français.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/damoko2004/statAfrikR}
  \item Report bugs at \url{https://github.com/damoko2004/statAfrikR/issues}
}

}
\author{
\strong{Maintainer}: Dickers Samoko \email{damoko2004@gmail.com}

}
\keyword{internal}
