% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{theme_ins}
\alias{theme_ins}
\title{Thème ggplot2 officiel INS}
\usage{
theme_ins(
  base_size = 11,
  base_family = "sans",
  couleur_fond = "white",
  grille = TRUE,
  grille_mineure = FALSE
)
}
\arguments{
\item{base_size}{numeric — Taille de base de la police. Défaut : 11.}

\item{base_family}{character — Famille de police. Défaut : "sans".}

\item{couleur_fond}{character — Couleur de fond du panneau. Défaut : "white".}

\item{grille}{logical — Afficher les lignes de grille. Défaut : TRUE.}

\item{grille_mineure}{logical — Afficher la grille mineure. Défaut : FALSE.}
}
\value{
Un objet \code{theme} ggplot2.
}
\description{
Applique un thème graphique professionnel adapté aux
publications officielles des Instituts Nationaux de Statistique africains.
Inspiré des chartes graphiques AFRISTAT/PARIS21.
}
\examples{
\dontrun{
  library(ggplot2)
  ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    theme_ins()
}
}
